Initial Commit of LightAuth2
-----------------------
What Was Wrong
-----------------------
I wanted to learn more go-lang and how to apply the 'clean architecture' to it. 

-----------------------
What I Did
-----------------------
I decided to re-implement the lightauth project in clean architecture terms and instead of json-rpc use json-REST. For more information see the main page documentation.

-----------------------
How To Verify Things
-----------------------
Build and Run:

   make test