Fixed undeploy script and added header values for host and version
-----------------------
What Was Wrong
-----------------------
The undeploy ansible script did not undeploy the application correctly and we wanted also to add header to show which worker handled the call. 

-----------------------
What We Did
-----------------------
Fixed the broken undeploy script and tested it. Wrote and added a couple of negroni middleware handlers to add a Header containing host name (X-Worker) and another one to contain the application version number (X-Worker-Version) 

-----------------------
How To Verify Things
-----------------------

1) Via ansible deploy the app and undeploy the app and check via 'ps -a' that the application is no longer running.
2) Make a http rest call and check the headers in the response and you should be able to see the two new headers.