Version 1.2 Added Authenticate via Claims
-----------------------

What Was Wrong
-----------------------
We want to add authenticate via claims so our 'authenticateme' project can authenticate a use via multiple mechanisms such as password or qrcode.

-----------------------
What We Did
-----------------------
We added a variation of the Authenticate method where is no password is present and 'claims' is then we aithenticate against the claims.

-----------------------
How To Verify Things
-----------------------
Run the unit test or execute via a Get POST the authenticate via claims variation.