Merged 3 apps into 1 by using cobra cli commands
-----------------------
What Was Wrong
-----------------------
I wanted to remove the need for three applications and instead replace them with one.

-----------------------
What I Did
-----------------------
I decided to re-implement the lightauth2 cli command parameters mechanism to one using
the cobra cli library where we have commands and parameters for 'serve', 'user' and 
'session'.  

-----------------------
How To Verify Things
-----------------------
Run the application in each of its three modes and check parameter passing works 
and the functionality on an API level works as previously.